+++ src/components/Wrappers/withUserRoles.js
@@ -5,6 +5,7 @@ import isEqual from 'lodash/isEqual';
 import { useCreateAuthUserKeycloak, useUserAffiliationRoles } from '../../hooks';
 import { KEYCLOAK_USER_EXISTANCE } from '../../constants';
 import { showErrorCallout } from '../../views/UserEdit/UserEditHelpers';
+import { isEmpty } from 'lodash';
 
 const withUserRoles = (WrappedComponent) => (props) => {
   const { okapi } = useStripes();
@@ -100,9 +101,13 @@ const withUserRoles = (WrappedComponent) => (props) => {
         break;
       case KEYCLOAK_USER_EXISTANCE.nonExist:
         // First, save changes to mod-users.
-        // If user decides to create a Keycloak user, then changes will be copied over from mod-users to mod-users-keycloak.
         await mutator.selUser.PUT(data);
-        setIsCreateKeycloakUserConfirmationOpen(true);
+
+        // Only prompt and create Keycloak user if assigning roles.
+        // If user confirms, then changes will be copied over from mod-users to mod-users-keycloak.
+        if (!isEqual(assignedRoleIds, initialAssignedRoleIds)) {
+          setIsCreateKeycloakUserConfirmationOpen(true);
+        }
         break;
       default:
         break;