+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 * Revert migration of custom fields title from mod-configuration to mod-settings. Ref UIU-3441.
 * *BREAKING* Display custom fields in different accordions depending on the selected section. Refs UIU-3131.
 * *BREAKING* Update the user record view page to display custom fields under the "User information", "Extended information", "Contact information", or default "Custom fields" accordions as configured in settings. Refs UIU-3129.
+* Only show Keycloak user record confirmation when assigning roles to a non-Keycloak record. Refs UIU-3426.
 
 ## [12.1.7] (https://github.com/folio-org/ui-users/tree/v12.1.7) (2025-06-03)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v12.1.6...v12.1.7)